Prepair for HW decoder
